refactor: extract getNormalizedMemoryGenerationState into memoryPromptPresets
